Move disjunction check to grounding level (#419)
* is_rule_with_builtin

* add test

* more readable code

* move disjunction check to grounding
