+++ neurolang/probabilistic/cplogic/program.py
@@ -4,7 +4,7 @@ from ...datalog import DatalogProgram
 from ...exceptions import ForbiddenDisjunctionError
 from ...expression_pattern_matching import add_match
 from ...expression_walker import ExpressionWalker, PatternWalker
-from ...expressions import Constant, FunctionApplication, Symbol
+from ...expressions import Constant, Symbol
 from ...logic import Implication, Union
 from ..exceptions import MalformedProbabilisticTupleError
 from ..expression_processing import (
@@ -188,21 +188,6 @@ class CPLogicMixin(PatternWalker):
             self.symbol_table[pred_symb], [expression]
         )
         return expression
-
-    @add_match(
-        Implication(FunctionApplication, ...),
-        lambda exp: (
-            exp.antecedent
-            != Constant[bool](True, auto_infer_type=False, verify_type=False)
-        ),
-    )
-    def prevent_intensional_disjunction(self, rule):
-        pred_symb = rule.consequent.functor
-        if pred_symb in self.symbol_table:
-            raise ForbiddenDisjunctionError(
-                "CP-Logic programs do not support disjunctions"
-            )
-        return self.statement_intensional(rule)
 
 
 class CPLogicProgram(CPLogicMixin, DatalogProgram, ExpressionWalker):
