+++ neurolang/probabilistic/cplogic/tests/test_cplogic_program.py
@@ -182,12 +182,3 @@ def test_add_probchoice_does_not_sum_to_one():
         cpl.add_probabilistic_choice_from_tuples(
             P, probchoice_as_tuples_iterable
         )
-
-
-def test_forbidden_disjunction():
-    rule_a = Implication(P(x), Q(x))
-    rule_b = Implication(P(y), Z(y))
-    code = Union((rule_a, rule_b))
-    cpl = CPLogicProgram()
-    with pytest.raises(ForbiddenDisjunctionError):
-        cpl.walk(code)
